python/pyproject.toml
[project]
name = "iceberg-compaction"
version = "0.1.0"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = []

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"
module-name = "iceberg_compaction._internal"
features = ["pyo3/extension-module"]

[tool.uv]
cache-keys = [{file = "pyproject.toml"}, {file = "Cargo.toml"}, {file = "**/*.rs"}]
package = false

[dependency-groups]
dev = ["maturin>=1.0,<2.0", "pytest>=8.0", "pyiceberg>=0.9.0", "google-auth>=2.0", "pyarrow"]
